+++ addons_custom/dz23_whatsapp/models/__init__.py
@@ -1,5 +1,6 @@
 from . import (
     business_action,
+    conversation,
     message_event,
     message_inbox,
     message_media,